versions/GUI Version/requirements.txt
moviepy
pillow
numpy==1.18.0